backend/requirements.txt
ultralytics
fastapi==0.60.1
aiofiles==0.4.0
python-multipart>=0.0.5

Cython
matplotlib==3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
tensorboard>=2.2
# torch>=1.7.0
# torchvision>=0.8.1
tqdm>=4.41.0

seaborn>=0.11.0
pandas
protobuf==3.20.*
pydantic